requirements.txt
# Core Framework
streamlit>=1.29.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Azure Services
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# LLM Providers
openai>=1.6.0
anthropic>=0.39.0
httpx>=0.27.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3

# RAG / Vector Search (optional, disable with RAG_ENABLED=false)
pgvector>=0.2.4

